refactor: replace Int id type with Long

+++ app/src/main/java/com/elliemoritz/coinbook/data/dao/AlarmsDao.kt
@@ -13,14 +13,14 @@ interface AlarmsDao {
     suspend fun getAlarmsList(): List<AlarmDbModel>
 
     @Query("SELECT * FROM alarms WHERE id = :id LIMIT 1")
-    suspend fun getAlarm(id: Int): AlarmDbModel
+    suspend fun getAlarm(id: Long): AlarmDbModel
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun addAlarm(alarm: AlarmDbModel)
 
     @Query("DELETE FROM alarms WHERE id = :id")
-    suspend fun removeAlarm(id: Int)
+    suspend fun removeAlarm(id: Long)
 
     @Query("SELECT COUNT(*) FROM alarms")
-    suspend fun getAlarmsCount(): Int
+    suspend fun getAlarmsCount(): Long
 }
