+++ app/src/main/java/com/elliemoritz/coinbook/data/dao/LimitsDao.kt
@@ -13,17 +13,17 @@ interface LimitsDao {
     suspend fun getLimitsList(): List<LimitDbModel>
 
     @Query("SELECT * FROM limits WHERE id = :id LIMIT 1")
-    suspend fun getLimit(id: Int): LimitDbModel
+    suspend fun getLimit(id: Long): LimitDbModel
 
     @Query("SELECT * FROM limits WHERE categoryId = :categoryId")
-    suspend fun getLimitByCategoryId(categoryId: Int): LimitDbModel?
+    suspend fun getLimitByCategoryId(categoryId: Long): LimitDbModel?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun addLimit(limit: LimitDbModel)
 
     @Query("DELETE FROM limits WHERE id = :id")
-    suspend fun removeLimit(id: Int)
+    suspend fun removeLimit(id: Long)
 
     @Query("SELECT COUNT(*) FROM limits")
-    suspend fun getLimitsCount(): Int
+    suspend fun getLimitsCount(): Long
 }
