+++ app/src/main/java/com/elliemoritz/coinbook/data/dbModels/LimitDbModel.kt
@@ -6,9 +6,9 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "limits")
 data class LimitDbModel(
     @PrimaryKey(autoGenerate = true)
-    val id: Int,
+    val id: Long,
     val limitAmount: Int,
     val realAmount: Int,
-    val categoryId: Int,
+    val categoryId: Long,
     val categoryName: String
 )
