+++ app/src/main/java/com/elliemoritz/coinbook/domain/repositories/ExpensesRepository.kt
@@ -7,14 +7,14 @@ interface ExpensesRepository {
 
     fun getExpensesList(): Flow<List<Expense>>
     fun getExpensesListForMonth(): Flow<List<Expense>>
-    fun getCategoryExpensesListForMonth(categoryId: Int): Flow<List<Expense>>
+    fun getCategoryExpensesListForMonth(categoryId: Long): Flow<List<Expense>>
     suspend fun removeAllExpenses()
 
-    fun getExpense(id: Int): Flow<Expense>
+    fun getExpense(id: Long): Flow<Expense>
     suspend fun addExpense(operation: Expense)
     suspend fun editExpense(operation: Expense)
     suspend fun removeExpense(operation: Expense)
 
     fun getTotalExpensesAmountForMonth(): Flow<Int>
-    fun getTotalExpensesAmountByCategoryForMonth(categoryId: Int): Flow<Int>
+    fun getTotalExpensesAmountByCategoryForMonth(categoryId: Long): Flow<Int>
 }
